fix(challenges): move creator-block helpers into challenge-block.service instead
challenge-visibility.ts is a pure-logic leaf with no test-time deps; pulling
user-preferences.service into it drags in the Redis cache graph
(user-preferences.service -> redis/caches.ts -> orchestrator/models.ts ->
redis/resource-data.redis.ts) and hits a circular-import TDZ crash. That
silently zeroed out three daily-challenge suites (37 tests) instead of
failing loudly, since vitest reports an empty suite as passing.

Move getChallengeExcludedUserIds/challengeCreatorBlockSql into a new
challenge-block.service.ts instead, alongside the other *.service.ts
modules that already depend on user-preferences.service. Revert
challenge-visibility.ts to its pre-move state.
